+++ modules/list_books.rb
@@ -1,0 +1,9 @@
+module ListBooks
+  def list_books
+    puts 'List of Books:'
+    @books.each do |book|
+      puts "Title: #{book.title.capitalize}, Author: #{book.author.capitalize}"
+    end
+    puts "\n"
+  end
+end
